+++ SpeedyWeatherInternals/src/SpeedyWeatherInternals.jl
@@ -1,0 +1,13 @@
+#= SpeedyWeatherInternals
+
+Common (internal) functions for SpeedyWeather.jl and related packages 
+like RingGrids.jl, LowerTriangularArrays.jl and SpeedyTransforms.jl =#
+module SpeedyWeatherInternals
+
+    # ARCHITECTURES (Device handling)
+    include("Architectures/Architectures.jl")
+
+    # UTILS (Kernel launching and various utilities)
+    include("Utils/Utils.jl")
+
+end