+++ src/models/barotropic.jl
@@ -80,6 +80,7 @@ function initialize!(model::Barotropic; time::DateTime = DEFAULT_DATE)
         SpectralGrid with nlayers=$(spectral_grid.nlayers) provided."
 
     # initialize components
+    initialize!(model.geometry, model)
     initialize!(model.time_stepping, model)
     initialize!(model.coriolis, model)
     initialize!(model.forcing, model)