+++ src/physics/large_scale_condensation.jl
@@ -5,13 +5,25 @@ large_scale_condensation!(::ColumnVariables, ::Nothing, ::PrimitiveEquation) = n
 
 export ImplicitCondensation
 """
-Large scale condensation as with implicit precipitation.
+Large-scale condensation with implicit time stepping.
 $(TYPEDFIELDS)"""
-@kwdef struct ImplicitCondensation{NF<:AbstractFloat} <: AbstractCondensation
-    "Relative humidity threshold [1 = 100%] to trigger condensation"
-    relative_humidity_threshold::NF = 1
+@kwdef mutable struct ImplicitCondensation{NF<:AbstractFloat} <: AbstractCondensation
+    "[OPTION] Relative humidity threshold [1 = 100%] to trigger condensation"
+    relative_humidity_threshold::NF = 0.95
 
-    "Time scale in multiples of time step Δt, the larger the less immediate"
+    "[OPTION] Reevaporation efficiency [1/(kg/kg)], 0 for no reevaporation"
+    reevaporation::NF = 30
+
+    "[OPTION] Convert rain below freezing to snow?"
+    snow::Bool = true
+
+    "[OPTION] Freezing temperature for snow fall [K]"
+    freezing_threshold::NF = 263
+
+    "[OPTION] Melting temperature for snow fall [K]"
+    melting_threshold::NF = 278
+
+    "[OPTION] Time scale in multiples of time step Δt, the larger the less immediate"
     time_scale::NF = 3
 end
 
@@ -33,18 +45,18 @@ function large_scale_condensation!(
     column::ColumnVariables,
     model::PrimitiveWet,
 )
-    #TODO not needed for no condensation
-    saturation_humidity!(column, model.clausius_clapeyron)
+    # dispatch by large scale condensation type
     large_scale_condensation!(column, model.large_scale_condensation, model)
 end
 
 # function barrier for ImplicitCondensation to unpack model
 function large_scale_condensation!( 
     column::ColumnVariables,
-    scheme::ImplicitCondensation,
+    condensation::ImplicitCondensation,
     model::PrimitiveWet,
 )
-    large_scale_condensation!(column, scheme,
+    saturation_humidity!(column, model.clausius_clapeyron)
+    large_scale_condensation!(column, condensation,
         model.clausius_clapeyron, model.geometry, model.planet, model.atmosphere, model.time_stepping)
 end
 
@@ -56,7 +68,7 @@ and temperature from latent heat release and integrates the
 large-scale precipitation vertically for output."""
 function large_scale_condensation!(
     column::ColumnVariables,
-    scheme::ImplicitCondensation,
+    condensation::ImplicitCondensation,
     clausius_clapeyron::AbstractClausiusClapeyron,
     geometry::Geometry,
     planet::AbstractPlanet,
@@ -67,47 +79,104 @@ function large_scale_condensation!(
     (; pres, temp, humid) = column          # prognostic vars (from previous time step for numerical stability)
     (; temp_tend, humid_tend) = column      # tendencies to write into
     (; sat_humid) = column                  # intermediate variable, calculated in thermodynamics!
-    
-    # precompute scaling constant for precipitation output
+
+    # precompute scaling constants to minimize divisions (used to convert between humidity [kg/kg] and precipitation [m])
     pₛ = pres[end]                          # surface pressure
     (; Δt_sec) = time_stepping
-    Δσ = geometry.σ_levels_thick
-    pₛΔt_gρ = (pₛ * Δt_sec)/(planet.gravity * atmosphere.water_density)
+    Δσ = geometry.σ_levels_thick            # layer thickness in sigma coordinates
+    pₛ_gρ = pₛ/(planet.gravity * atmosphere.water_density)
 
-    (; Lᵥ, cₚ, Lᵥ_Rᵥ) = clausius_clapeyron
-    Lᵥ_cₚ = Lᵥ/cₚ                           # latent heat of vaporization over heat capacity
-    (; time_scale, relative_humidity_threshold) = scheme
+    # thermodynamics
+    Lᵥ = clausius_clapeyron.latent_heat_condensation    # latent heat of vaporization
+    Lᵢ = clausius_clapeyron.latent_heat_fusion          # latent heat of fusion
+    cₚ = clausius_clapeyron.heat_capacity               # heat capacity
+    Rᵥ = clausius_clapeyron.R_vapour                    # gas constant for water vapour
+    Lᵥ_cₚ = Lᵥ / cₚ
+    Lᵢ_cₚ = Lᵢ / cₚ
+    cₚ_Lᵢ = cₚ / Lᵢ
+
+    (; time_scale, relative_humidity_threshold, freezing_threshold, melting_threshold) = condensation
+    let_it_snow = condensation.snow                     # flag to switch snow on/off
+    rain_flux_down = zero(eltype(column))               # start with zero rain/snow flux at top of atmosphere
+    snow_flux_down = zero(eltype(column))               # both have units of [m/s] (precipitation)
 
     @inbounds for k in eachindex(column)
-        if humid[k] > sat_humid[k]*relative_humidity_threshold
 
-            # tendency for Implicit humid = sat_humid, divide by leapfrog time step below
-            δq = sat_humid[k] * relative_humidity_threshold - humid[k]
+        # Condensation from humidity in this layer (for a negative humidity tendency)
+        # relative to threshold that can be <100%, e.g. 95%
+        δq_cond = sat_humid[k] * relative_humidity_threshold - humid[k]
+        
+        # skip if no condensation has occurred yet in this layer or above
+        if δq_cond < 0 || snow_flux_down > 0 || rain_flux_down > 0
 
+            # 0. convert between humidity tendency [kg/kg/s] and precipitation amount [m] or rate [m/s]
+            Δp_gρ = Δσ[k] * pₛ_gρ                           # pressure thickness of layer Δp times 1/g/ρ [m]
+            Δp_Δtgρ = Δp_gρ/Δt_sec                          # pressure thickness of layer Δp times 1/Δt/g/ρ [m/s]
+            Δtgρ_Δp = inv(Δp_Δtgρ)                          # [s/m]
+
+            # 1. Melting of snow from layer above
+            δT_melt = max(temp[k] - melting_threshold, 0)   # only if temperature above melting threshold
+            E_avail = cₚ_Lᵢ * δT_melt                       # available energy [1] in multiples of latent heat [J/kg of air] for melting
+            melt_rate_max = E_avail * Δp_Δtgρ               # max rate of snow melt [m/s]
+            melt_rate = min(snow_flux_down, melt_rate_max)  # cap to snow flux [m/s], don't melt more than available
+            snow_flux_down -= melt_rate                     # move melted snow to rain
+            rain_flux_down += melt_rate                     # this can evaporate now too
+            δq_melt = melt_rate * Δtgρ_Δp                   # convert back to humidity increase over timestep [kg/kg]
+            δT = -Lᵢ_cₚ * δq_melt                           # just to calculate the latent heat required (explicit in time)
+                                                            # don't use δq_melt as humidity tendency as it's only snow->rain
+
+            # 2. Reevaporation and condensation
+            dq = sat_humid[k] - humid[k]                    # reevaporation relative to 100% humidity
+            r = condensation.reevaporation * max(dq, 0)     # reevaporation efficiency [1], scale linearly with dryness
+            rain_evaporated = min(r, 1) * rain_flux_down    # [m/s], min(r, 1) to not evaporate more than available
+            rain_flux_down -= rain_evaporated               # remove reevaporated rain
+            δq_evap = rain_evaporated * Δtgρ_Δp             # convert to humidity tendency over timestep [kg/kg]
+            δq = min(0, δq_cond) + δq_evap                  # [kg/kg] sum with condensation (negative) and evaporation (positive)
+                                                            # division by timestep for tendency below
+
+            # Solve for melting of snow, condensation, reevaporation (and possibly sublimation) implicitly in time
             # implicit correction, Frierson et al. 2006 eq. (21)
-            dqsat_dT = sat_humid[k] * relative_humidity_threshold * Lᵥ_Rᵥ/temp[k]^2       # derivative of qsat wrt to temp
-            δq /= ((1 + Lᵥ_cₚ*dqsat_dT) * time_scale*Δt_sec) 
-
-            # latent heat release for enthalpy conservation
-            δT = -Lᵥ_cₚ * δq
+            # derivative of qsat wrt to temp
+            T = temp[k] + Δt_sec*δT             # use temperature minus latent heat for melting in gradient
+            dqsat_dT = sat_humid[k] * relative_humidity_threshold * Lᵥ_cₚ/(Rᵥ*T^2)
+            δq /= ((1 + Lᵥ_cₚ*dqsat_dT) * time_scale*Δt_sec)
+            δT = -Lᵥ_cₚ * δq                    # latent heat release for enthalpy conservation
 
             # If there is large-scale condensation at a level higher (i.e. smaller k) than
             # the cloud-top previously diagnosed due to convection, then increase the cloud-top
             # Fortran SPEEDY documentation Page 7 (last sentence)
             column.cloud_top = min(column.cloud_top, k)
-    
-            # 2. Precipitation due to large-scale condensation [kg/m²/s] /ρ for [m/s]
-            # += for vertical integral
-            precip = Δσ[k] * pₛΔt_gρ * -δq          # precipitation [m] on layer k, Formula 4
-            column.precip_large_scale += precip     # integrate vertically, Formula 25, unit [m]
+
+            # 2. Precipitation (rain) due to large-scale condensation [kg/m²/s] /ρ for [m/s]
+            δq_rain = min(0, δq)                # precipitation only for negative humidity tendency
+            rain = Δp_gρ * -δq_rain             # precipitation rain [m/s] on this layer k, Formula 4
+            snow = zero(rain)                   # start with zero snow but potentially swap below
+
+            # decide whether to turn precip into snow (all rain freezes to snow)
+            rain, snow = let_it_snow && temp[k] < freezing_threshold ? (snow, rain) : (rain, snow)
+            rain_flux_down += rain              # accumulate into downward fluxes [m/s] (used in layer below)
+            snow_flux_down += snow              # accumulate into downward fluxes [m/s] (used in layer below)
+
+            # latent heat release when freezing for enthalpy conservation
+            δT -= (snow > 0) * Lᵢ_cₚ * δq_rain  # snow for negative δq (condensation then freezing)
 
             # only accumulate into humid_tend now to allow humid_tend != 0 before this scheme is called
-            humid_tend[k] += δq
-            temp_tend[k] += δT
+            humid_tend[k] += δq                 # [kg/kg/s]
+            temp_tend[k] += δT                  # [K/s]
         end
     end
 
-    # convert to rain rate [m/s]
-    column.precip_rate_large_scale = column.precip_large_scale / Δt_sec
+    # avoid negative precipitation from rounding errors 
+    rain_flux_down = max(rain_flux_down, 0)
+    snow_flux_down = max(snow_flux_down, 0)
+
+    # precipitation from rain/snow [m] during time step whatever is fluxed out 
+    column.rain_large_scale = Δt_sec*rain_flux_down
+	column.snow_large_scale = Δt_sec*snow_flux_down
+
+    # and the rain/snow fall rate [m/s]
+    column.rain_rate_large_scale = rain_flux_down
+    column.snow_rate_large_scale = snow_flux_down
+
     return nothing
 end